ADR-0063 impl: HtmlToMarkdown primitive + 3 caller swaps
Promote the HTML→Markdown conversion (previously buried in
MarkdownContentExtractor) to a public static primitive under
WebReaper.Core.Markdown. The adapter shrinks to a ~20-line shell over
the primitive; three of four callers swap to call the primitive
directly, resolving the structural awkwardness where they had to
instantiate and call the *adapter* for what is a pure *function*.

New
- WebReaper/Core/Markdown/HtmlToMarkdown.cs — public static class with
  Convert(html) → string and ExtractMainContent(html) → MainContent.
  AOT-clean (synchronous, no reflection, AngleSharp HtmlParser sync
  path instead of BrowsingContext.OpenAsync). The heuristic
  (article → main → [role=main] → body), strip list, GFM render, and
  title resolution all move here verbatim.
- WebReaper/Core/Markdown/MainContent.cs — public record
  MainContent(Title, Markdown).
- WebReaper.Tests/WebReaper.UnitTests/HtmlToMarkdownTests.cs — 31
  pure-function tests: heuristic priority (article > main > role=main
  > body), strip list (nav/footer/aside/script/style/noscript/hidden/
  aria-hidden), GFM grammar (heading levels, paragraphs, lists,
  links, images, code blocks, blockquotes, tables), title resolution
  (h1-inside-main > head/title, post-strip), edge cases (empty body,
  null input, whitespace normalisation), Convert overload pure-
  function shape.

Shrunk
- WebReaper/Core/Parser/Concrete/MarkdownContentExtractor.cs — 593 →
  64 lines. Adapter shell calls HtmlToMarkdown.ExtractMainContent then
  wraps as JsonObject {title, markdown}. Stays public (the
  AsMarkdown() seed terminal wires it into ScraperConfig).

Caller swaps
- WebReaper.AI/LlmContentExtractor.cs — drop the
  MarkdownContentExtractor field; PreCleanToMarkdownAsync becomes
  synchronous PreCleanToMarkdown calling
  HtmlToMarkdown.ExtractMainContent directly.
- WebReaper/Core/Agent/Concrete/AgentEngine.cs — drop the static
  MarkdownView field; the brain's page-Markdown view now calls
  HtmlToMarkdown.Convert(pageHtml). The try/catch around rendering
  stays (a corrupt page might break the parser).
- WebReaper/Processing/Concrete/ChangeTrackingProcessor.cs — drop
  the MarkdownContentExtractor field; hash input is now
  HtmlToMarkdown.Convert(context.Html) directly (no JsonObject
  construct-and-discard).

Tests added
- ChangeTrackingProcessorTests.cs — pin template-noise hash stability:
  a footer timestamp that rotates between visits must NOT flip the
  status (the strip list removes <footer> before hashing).

ADR-0063 status flipped Proposed → Accepted — implemented (2026-05-24).

Build: 0 errors. Unit tests: 319 pass. AI tests: 53 pass. Agent tests:
27 pass (all transparent to the swap).

NOTE for the integrator: ADR-0059 (LlmCall<T> consolidation) and
ADR-0063 both touch WebReaper.AI/LlmContentExtractor.cs. This branch
makes the *minimum* surface change there — swapping the pre-clean
indirection (await _markdown.ExtractAsync(...).["markdown"]) for the
direct primitive call (HtmlToMarkdown.ExtractMainContent(...).Markdown).
LlmSelectorRepairer.cs was left untouched (still uses the
MarkdownContentExtractor shell, which works correctly) — ADR-0059's
mechanism refactor should handle that file in one pass. At integration
time, if 0059 reaches LlmContentExtractor.PreCleanToMarkdown, prefer
the HtmlToMarkdown.ExtractMainContent direct-call shape from this
branch.
